unibookswap_api_bkup/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
commonmark==0.9.1
cryptography==44.0.3
defusedxml==0.7.1
deprecation==2.1.0
dj-rest-auth==5.0.2
Django==5.2
django-allauth==0.61.1
django-channels==0.7.0
django-cors-headers==4.7.0
django-email-verification==0.3.3
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
idna==3.10
mutagen==1.47.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==25.0
pyasn1==0.4.8
pycparser==2.22
Pygments==2.19.1
PyJWT==1.7.1
python-jose==3.4.0
python3-openid==3.2.0
redis==6.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==12.6.0
rsa==4.9.1
sentry-sdk==1.45.1
six==1.17.0
spotify_dl==8.9.0
spotipy==2.25.1
sqlparse==0.5.3
urllib3==1.26.20
validators==0.34.0
yt-dlp==2025.4.30